Улучшена валидация, decimal-расчёты и профили лазера
Добавлена строгая валидация пользовательских данных во всех основных полях ввода (толщина, количество, размеры, масса, резка, гибка, сварка) с помощью IDataErrorInfo и ValidationRule, ошибки отображаются под полями. Вся бизнес-логика расчётов и модели переведены на decimal для точности. В SettingsWindow реализовано редактирование профилей резки лазера. Добавлен интерфейс ICalculationService и модульные тесты для расчётов. Внесены улучшения в XAML: шаблоны ошибок, единицы измерения, валидация числовых полей. Исправлена работа с decimal в базе и моделях. Улучшен UI и поддержка новых зависимостей.

+++ MetalCalcWPF/Models/MaterialType.cs
@@ -9,7 +9,7 @@ namespace MetalCalcWPF.Models
 
         public string Name { get; set; }      // "Черная сталь (Ст3)", "Нержавейка"
         public double Density { get; set; }   // Плотность (г/см3). Для стали 7.85
-        public double BasePricePerKg { get; set; } // Базовая цена (если фиксированная)
+        public decimal BasePricePerKg { get; set; } // Базовая цена (если фиксированная)
 
         // Переопределение метода ToString, чтобы в выпадающем списке писалось название
         public override string ToString()
